17th commit : 전체 목록 지도로 띄우기 완료

+++ lib/tabs/tab2_addition/AddSpot.dart
@@ -314,7 +314,7 @@ class _MapBottomSheetState extends State<MapBottomSheet> {
                   ),
                   onTap: () {
                     // 호텔 선택 시 콜백 호출
-                    widget.onSpotSelected(spot[0], spot[1].toDouble(), spot[2].toDouble(), spot[3], spot[4], spot[5]);
+                    widget.onSpotSelected(spot[0], spot[1].toDouble(), spot[2].toDouble(), spot[3], spot[4].toDouble(), spot[5]);
                   },
                 ),
               );
